finihing

+++ client/src/fetches/heroFetch.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const URL = 'http://localhost:3000/heroes';
 const token = localStorage.getItem('access_token');
@@ -52,7 +53,7 @@ const getHeroStats = async (id, cb) => {
 }
 
 const searchHero = async (cb) => {
-    try{
+    try {
 
     } catch (e) {
         console.log(e);
@@ -65,7 +66,7 @@ const register = async hero => {
             method: 'POST',
             url: URL + '/registration',
             data: hero,
-            headers: {          
+            headers: {
                 'Content-Type': 'multipart/form-data'
             }
         });
@@ -83,7 +84,7 @@ const addHero = async hero => {
             url: URL + '/add-hero',
             data: hero,
             headers: {
-                'Accept': 'application/json',                
+                'Accept': 'application/json',
                 'Content-Type': 'multipart/form-data'
             }
         });
@@ -128,10 +129,28 @@ const updateHero = async (id, hero) => {
 
 const removeHero = async id => {
     try {
-        let result = await axios({
-            method: 'DELETE',
-            url: URL + '/delete-hero/' + id
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, kill it!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                let result = await axios({
+                    method: 'DELETE',
+                    url: URL + '/delete-hero/' + id
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+            }
+            window.location.reload(true)
+        })
 
     } catch (e) {
         console.log(e);
